Made a separated header for NonterminalTokenKeyTable
