+++ include/CYK_Algorithm.h
@@ -4,6 +4,6 @@
 
 #include <string_view>
 
-namespace fl::cyk {
+namespace fl::algo::cyk {
     bool isRecognized(const std::string& text, const Grammar& g);
 }  // namespace fl::cyk
